access des role pour produits

+++ src/Controller/ProduitController.php
@@ -25,11 +25,11 @@ class ProduitController extends AbstractController
 
     #[Route('/new', name: 'app_produit_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
-    { {   // Vérification des droits d’accès
-            if (!$this->isGranted('ROLE_PROSPECT')) {
-                throw $this->createAccessDeniedException('Accès refusé');
-            }
+    {
+        if (!$this->isGranted('ROLE_PROSPECT') && !$this->isGranted('ROLE_ADMIN') && !$this->isGranted('ROLE_MARAICHER')) {
+            throw $this->createAccessDeniedException('Accès refusé');
         }
+
         $produit = new Produit();
         $form = $this->createForm(Produit1Type::class, $produit);
         $form->handleRequest($request);
@@ -81,6 +81,9 @@ class ProduitController extends AbstractController
     #[Route('/{id}/edit', name: 'app_produit_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Produit $produit, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->isGranted('ROLE_PROSPECT') && !$this->isGranted('ROLE_ADMIN') && !$this->isGranted('ROLE_MARAICHER')) {
+            throw $this->createAccessDeniedException('Accès refusé');
+        }
         $form = $this->createForm(Produit1Type::class, $produit);
         $form->handleRequest($request);
 
@@ -102,6 +105,9 @@ class ProduitController extends AbstractController
     #[Route('/{id}', name: 'app_produit_delete', methods: ['POST'])]
     public function delete(Request $request, Produit $produit, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN') && !$this->isGranted('ROLE_MARAICHER')) {
+            throw $this->createAccessDeniedException('Accès refusé');
+        }
         if ($this->isCsrfTokenValid('delete' . $produit->getId(), $request->request->get('_token'))) {
             $entityManager->remove($produit);
             $entityManager->flush();
